Refactor PDF generation into a separate module with a configurable template.
This commit separates the PDF generation logic from the main Flask application file (`app.py`) into a dedicated module (`pdf_generator.py`).

Key changes:
- A new `pdf_generator.py` module now contains all the logic for creating the PDF, including HTML parsing, section reordering, and TOC generation.
- The PDF's HTML structure and CSS have been externalized into a Jinja2 template (`templates/pdf_template.html`).
- A new `config.py` file has been added to specify the path to the PDF template, making it easily configurable.
- The `download_pdf` route in `app.py` has been simplified to act as an orchestrator, calling the new module.
- The `Jinja2` dependency has been added to `requirements.txt`.

This refactoring improves modularity, maintainability, and makes it easier to customize the PDF's appearance without modifying the application's core logic.
