+++ config.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+"""
+Configuration file for the application.
+"""
+
+# Path to the PDF template file.
+# This template is used by the PDF generator to create the final PDF.
+PDF_TEMPLATE_PATH = 'templates/pdf_template.html'
