+++ requirements.txt
@@ -12,3 +12,4 @@ langchain-mistralai
 requests
 json5
 better-profanity
+Jinja2
