+++ templates/pdf_template.html
@@ -1,0 +1,96 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>{{ title }}</title>
+    <style>
+        @import url('https://fonts.googleapis.com/css2?family=Merriweather:wght@400;700&family=Roboto:wght@400;700&display=swap');
+
+        /* General Body and Font Styles */
+        body { font-family: 'Merriweather', serif; line-height: 1.6; color: #333; }
+        h1, h2, h3, h4, h5, h6 { font-family: 'Roboto', sans-serif; font-weight: 700; }
+        h2 {
+            border-bottom: 2px solid #cccccc;
+            padding-bottom: 10px;
+            font-size: 24pt;
+            string-set: current_section content();
+            page-break-after: avoid;
+        }
+        h3 {
+            font-size: 18pt;
+            border-bottom: 1px solid #eeeeee;
+            padding-bottom: 5px;
+            page-break-after: avoid;
+        }
+        pre { background-color: #f5f5f5; padding: 1em; border-radius: 4px; white-space: pre-wrap; word-wrap: break-word; font-family: 'Courier New', Courier, monospace; }
+        table { border-collapse: collapse; width: 100%; margin-top: 1em;}
+        th, td { border: 1px solid #ddd; padding: 8px; text-align: left; }
+        th { background-color: #f2f2f2; }
+
+        /* Page Definitions */
+        @page {
+            size: A4;
+            margin: 2cm;
+            @bottom-left {
+                content: string(current_section, first);
+                font-size: 9pt;
+                color: #888;
+            }
+            @bottom-right {
+                content: "Page " counter(page);
+                font-size: 9pt;
+                color: #888;
+            }
+        }
+        @page:first {
+            margin: 1cm;
+            @bottom-center {
+                content: "Generated by the \"Scenario Generator X Papoulos\"";
+                font-size: 9pt;
+                color: #888;
+                vertical-align: top;
+            }
+        }
+
+        /* Cover Page */
+        .cover-page {
+            page-break-after: always;
+            width: 100%;
+            height: 90vh;
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
+            text-align: center;
+        }
+        .cover-title {
+            font-size: 40pt;
+        }
+
+        /* Table of Contents */
+        #toc {
+            page-break-after: always;
+        }
+        #toc h2 {
+            text-align: center;
+            border-bottom: 2px solid #cccccc;
+            padding-bottom: 10px;
+            string-set: current_section "Table des Matières";
+        }
+        #toc ul { list-style-type: none; padding: 0; }
+        #toc li a { text-decoration: none; color: #333; display: block; padding: 5px 0; }
+        #toc li a::after { content: leader('.') target-counter(attr(href), page); }
+
+        /* Other Sections */
+        .new-page { page-break-after: always; }
+    </style>
+</head>
+<body>
+    <div class="cover-page">
+        <h1 class="cover-title">{{ title }}</h1>
+    </div>
+
+    {{ toc|safe }}
+
+    {{ content|safe }}
+</body>
+</html>
